增加hudi json写能力功能实现
