+++ pom.xml
@@ -42,8 +42,7 @@
 		<module>chunjun-assembly</module>
 		<module>chunjun-e2e</module>
 		<module>chunjun-local-test</module>
-        <module>chunjun-connectors/chunjun-connector-hudi</module>
-    </modules>
+	</modules>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -76,7 +75,7 @@
 		<fastjson.version>1.2.83</fastjson.version>
 		<hutool.version>5.8.4</hutool.version>
 		<protostuff.version>1.8.0</protostuff.version>
-		<avro.version>1.8.2</avro.version>
+		<avro.version>1.10.0</avro.version>
 		<codec.version>1.13</codec.version>
 		<jcommander.version>1.81</jcommander.version>
 		<p3c-pmd.version>1.3.0</p3c-pmd.version>
